introduce global context and migrate updater controller into
