+++ Easydict/App/AppDelegate.h
@@ -11,6 +11,4 @@
 
 @interface AppDelegate : NSObject <NSApplicationDelegate>
 
-@property (weak) IBOutlet SPUStandardUpdaterController *updaterController;
-
 @end
