+++ Easydict/NewApp/EasydictApp.swift
@@ -8,6 +8,7 @@
 
 import Sparkle
 import SwiftUI
+import Defaults
 
 @main
 enum EasydictCmpatibilityEntry {
@@ -21,28 +22,12 @@ enum EasydictCmpatibilityEntry {
     }
 }
 
-class SPUUpdaterHelper: NSObject, SPUUpdaterDelegate {
-    func feedURLString(for _: SPUUpdater) -> String? {
-        var feedURLString = "https://raw.githubusercontent.com/tisfeng/Easydict/main/appcast.xml"
-        #if DEBUG
-            feedURLString = "http://localhost:8000/appcast.xml"
-        #endif
-        return feedURLString
-    }
-}
-
-class SPUUserDriverHelper: NSObject, SPUStandardUserDriverDelegate {
-    var supportsGentleScheduledUpdateReminders: Bool {
-        true
-    }
-}
-
 struct EasydictApp: App {
     @NSApplicationDelegateAdaptor
-    var delegate: AppDelegate
+    private var delegate: AppDelegate
 
-    @AppStorage(kHideMenuBarIconKey)
-    private var hideMenuBar = false
+    @Default(.hideMenuBarIcon)
+    private var hideMenuBar
 
     private var menuBarImage: String {
         #if DEBUG
@@ -52,22 +37,10 @@ struct EasydictApp: App {
         #endif
     }
 
-    let userDriverHelper = SPUUserDriverHelper()
-    let upadterHelper = SPUUpdaterHelper()
-
-    private let updaterController: SPUStandardUpdaterController
-
-    init() {
-        // 参考 https://sparkle-project.org/documentation/programmatic-setup/
-        // If you want to start the updater manually, pass false to startingUpdater and call .startUpdater() later
-        // This is where you can also pass an updater delegate if you need one
-        updaterController = SPUStandardUpdaterController(startingUpdater: true, updaterDelegate: upadterHelper, userDriverDelegate: userDriverHelper)
-    }
-
     var body: some Scene {
         if #available(macOS 13, *) {
             MenuBarExtra(isInserted: $hideMenuBar.toggledValue) {
-                MenuItemView(updater: updaterController.updater)
+                MenuItemView()
             } label: {
                 Label {
                     Text("Easydict")
